requirements.txt

transformers
torch
pytesseract
Pillow
requests
